pyproject.toml
[project]
name = "aitaem"
version = "0.1.0"
description = "All Interesting Things Are Essentially Metrics - A Python library for generating data insights"
requires-python = ">=3.9"
dependencies = [
    "ibis-framework[duckdb]>=9.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
bigquery = [
    "ibis-framework[bigquery]>=9.0.0",
    "google-cloud-bigquery>=3.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
all = ["aitaem[bigquery,dev]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require real external resources)",
]
# By default, skip integration tests unless explicitly requested
addopts = "-m 'not integration'"

[tool.ruff]
line-length = 100
